// interfaces/filter-options.ts
export interface FilterOptions {
    installationStartDate?: Date;
    installationEndDate?: Date;
    installationDate?: Date;
    isSolarPanelInstalled?: boolean;
    startDate?: Date;
    endDate?: Date;
    method?: string;
    action?: string;
    numberOfConsumers?: number;
    city?: string;
    state?: string;
}